ANT-26: Finished MongoDB Integration
