+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = cabinet
-version = 2023.06.23.1
+version = 2023.06.25.1
 author = Tyler Woodfin
 author_email = feedback@tyler.cloud
 description = Easily manage data storage and logging across repos
