+++ src/cabinet/__init__.py
@@ -1,5 +1,4 @@
 """
 Allows for easy Cabinet import
 """
-# from .cabinet import Cabinet, Mail, main
-from .mongodb import Cabinet, Mail, main
+from .cabinet import Cabinet, Mail, main
